requirements.txt
python-telegram-bot==20.7

playwright==1.40.0

# HTTP запити (для mangalib API)
aiohttp==3.9.1

python-dotenv==1.0.0

filelock==3.13.1

# Опціонально — прихований режим якщо сайти починають блокувати
# playwright-stealth==1.0.6